35 Harvey Street
"We could see the other side of the road and the Park."

"Quite so. Then I do not think that I need to detain you longer. I shall communicate with you."

"Should you be fortunate enough to solve this problem," said our client, rising.

"I have solved it."

"Eh? What was that?"

"I say that I have solved it."