107 Glan Road
It was early in April in the year '83 that I woke one morning to find Sherlock Holmes standing,
fully dressed, by the side of my bed. He was a late riser, as a rule, and as the clock on the
mantelpiece showed me that it was only a quarter-past seven, I blinked up at him in some surprise,
and perhaps just a little resentment, for I was myself regular in my habits.

"Very sorry to knock you up, Watson," said he, "but it's the common lot this morning. Mrs. Hudson
has been knocked up, she retorted upon me, and I on you."

"What is it, then--a fire?"

"No; a client. It seems that a young lady has arrived in a considerable state of excitement, who
insists upon seeing me. She is waiting now in the sitting-room. Now, when young ladies wander about
the metropolis at this hour of the morning, and knock sleepy people up out of their beds, I presume
that it is something very pressing which they have to communicate. Should it prove to be an
interesting case, you would, I am sure, wish to follow it from the outset. I thought, at any rate,
that I should call you and give you the chance."