61 Down Close
It was a few weeks before my own marriage, during the days when I was still sharing rooms with
Holmes in Baker Street, that he came home from an afternoon stroll to find a letter on the table
waiting for him. I had remained indoors all day, for the weather had taken a sudden turn to rain,
with high autumnal winds, and the Jezail bullet which I had brought back in one of my limbs as a
relic of my Afghan campaign throbbed with dull persistence. With my body in one easy-chair and my
legs upon another, I had surrounded myself with a cloud of newspapers until at last, saturated with
the news of the day, I tossed them all aside and lay listless, watching the huge crest and monogram
upon the envelope upon the table and wondering lazily who my friend's noble correspondent could be.

"Here is a very fashionable epistle," I remarked as he entered. "Your morning letters, if I remember
right, were from a fish-monger and a tide-waiter."

"Yes, my correspondence has certainly the charm of variety," he answered, smiling, "and the humbler
are usually the more interesting. This looks like one of those unwelcome social summonses which call
upon a man either to be bored or to lie."

He broke the seal and glanced over the contents.