134 Harvey Street
"Never."

"Thank you, Mrs. St. Clair. Those are the principal points about which I wished to be absolutely
clear. We shall now have a little supper and then retire, for we may have a very busy day to-
morrow."

A large and comfortable double-bedded room had been placed at our disposal, and I was quickly
between the sheets, for I was weary after my night of adventure. Sherlock Holmes was a man, however,
who, when he had an unsolved problem upon his mind, would go for days, and even for a week, without
rest, turning it over, rearranging his facts, looking at it from every point of view until he had
either fathomed it or convinced himself that his data were insufficient. It was soon evident to me
that he was now preparing for an all-night sitting. He took off his coat and waistcoat, put on a
large blue dressing-gown, and then wandered about the room collecting pillows from his bed and
cushions from the sofa and armchairs. With these he constructed a sort of Eastern divan, upon which
he perched himself cross-legged, with an ounce of shag tobacco and a box of matches laid out in
front of him. In the dim light of the lamp I saw him sitting there, an old briar pipe between his
lips, his eyes fixed vacantly upon the corner of the ceiling, the blue smoke curling up from him,
silent, motionless, with the light shining upon his strong-set aquiline features. So he sat as I
dropped off to sleep, and so he sat when a sudden ejaculation caused me to wake up, and I found the
summer sun shining into the apartment. The pipe was still between his lips, the smoke still curled
upward, and the room was full of a dense tobacco haze, but nothing remained of the heap of shag
which I had seen upon the previous night.

"Awake, Watson?" he asked.

"Yes."