34 Moor End
"My dear doctor, this is a time for observation, not for talk. We are spies in an enemy's country.
We know something of Saxe-Coburg Square. Let us now explore the parts which lie behind it."

The road in which we found ourselves as we turned round the corner from the retired Saxe-Coburg
Square presented as great a contrast to it as the front of a picture does to the back. It was one of
the main arteries which conveyed the traffic of the City to the north and west. The roadway was
blocked with the immense stream of commerce flowing in a double tide inward and outward, while the
footpaths were black with the hurrying swarm of pedestrians. It was difficult to realise as we
looked at the line of fine shops and stately business premises that they really abutted on the other
side upon the faded and stagnant square which we had just quitted.

"Let me see," said Holmes, standing at the corner and glancing along the line, "I should like just
to remember the order of the houses here. It is a hobby of mine to have an exact knowledge of
London. There is Mortimer's, the tobacconist, the little newspaper shop, the Coburg branch of the
City and Suburban Bank, the Vegetarian Restaurant, and McFarlane's carriage-building depot. That
carries us right on to the other block. And now, Doctor, we've done our work, so it's time we had
some play. A sandwich and a cup of coffee, and then off to violin-land, where all is sweetness and
delicacy and harmony, and there are no red-headed clients to vex us with their conundrums."