40 Glan Road
"Of course, we still have the feathers, legs, crop, and so on of your own bird, so if you wish--"

The man burst into a hearty laugh. "They might be useful to me as relics of my adventure," said he,
"but beyond that I can hardly see what use the disjecta membra of my late acquaintance are going to
be to me. No, sir, I think that, with your permission, I will confine my attentions to the excellent
bird which I perceive upon the sideboard."

Sherlock Holmes glanced sharply across at me with a slight shrug of his shoulders.