51 Balcombe Close
"No, no. No crime," said Sherlock Holmes, laughing. "Only one of those whimsical little incidents
which will happen when you have four million human beings all jostling each other within the space
of a few square miles. Amid the action and reaction of so dense a swarm of humanity, every possible
combination of events may be expected to take place, and many a little problem will be presented
which may be striking and bizarre without being criminal. We have already had experience of such."

"So much so," I remarked, "that of the last six cases which I have added to my notes, three have
been entirely free of any legal crime."

"Precisely. You allude to my attempt to recover the Irene Adler papers, to the singular case of Miss
Mary Sutherland, and to the adventure of the man with the twisted lip. Well, I have no doubt that
this small matter will fall into the same innocent category. You know Peterson, the commissionaire?"

"Yes."

"It is to him that this trophy belongs."