89 Cliff Place
Dr. Grimesby Roylott's chamber was larger than that of his step-daughter, but was as plainly
furnished. A camp-bed, a small wooden shelf full of books, mostly of a technical character, an
armchair beside the bed, a plain wooden chair against the wall, a round table, and a large iron safe
were the principal things which met the eye. Holmes walked slowly round and examined each and all of
them with the keenest interest.

"What's in here?" he asked, tapping the safe.