131 Moor End
"That's the list of the folk from whom I buy. D'you see? Well, then, here on this page are the
country folk, and the numbers after their names are where their accounts are in the big ledger. Now,
then! You see this other page in red ink? Well, that is a list of my town suppliers. Now, look at
that third name. Just read it out to me."

"Mrs. Oakshott, 117, Brixton Road--249," read Holmes.

"Quite so. Now turn that up in the ledger."

Holmes turned to the page indicated. "Here you are, 'Mrs. Oakshott, 117, Brixton Road, egg and
poultry supplier.'"