91 Larch Walk
"By this time the whole house was astir, for I had raised my voice in my anger. Mary was the first
to rush into my room, and, at the sight of the coronet and of Arthur's face, she read the whole
story and, with a scream, fell down senseless on the ground. I sent the house-maid for the police
and put the investigation into their hands at once. When the inspector and a constable entered the
house, Arthur, who had stood sullenly with his arms folded, asked me whether it was my intention to
charge him with theft. I answered that it had ceased to be a private matter, but had become a public
one, since the ruined coronet was national property. I was determined that the law should have its
way in everything.

"'At least,' said he, 'you will not have me arrested at once. It would be to your advantage as well
as mine if I might leave the house for five minutes.'

"'That you may get away, or perhaps that you may conceal what you have stolen,' said I. And then,
realising the dreadful position in which I was placed, I implored him to remember that not only my
honour but that of one who was far greater than I was at stake; and that he threatened to raise a
scandal which would convulse the nation. He might avert it all if he would but tell me what he had
done with the three missing stones.

"'You may as well face the matter,' said I; 'you have been caught in the act, and no confession
could make your guilt more heinous. If you but make such reparation as is in your power, by telling
us where the beryls are, all shall be forgiven and forgotten.'

"'Keep your forgiveness for those who ask for it,' he answered, turning away from me with a sneer. I
saw that he was too hardened for any words of mine to influence him. There was but one way for it. I
called in the inspector and gave him into custody. A search was made at once not only of his person
but of his room and of every portion of the house where he could possibly have concealed the gems;
but no trace of them could be found, nor would the wretched boy open his mouth for all our
persuasions and our threats. This morning he was removed to a cell, and I, after going through all
the police formalities, have hurried round to you to implore you to use your skill in unravelling
the matter. The police have openly confessed that they can at present make nothing of it. You may go
to any expense which you think necessary. I have already offered a reward of £1000. My God, what
shall I do! I have lost my honour, my gems, and my son in one night. Oh, what shall I do!"