103 Bowler Street
"Here is an advertisement which will interest you," said he. "It appeared in all the papers about a
year ago. Listen to this:

"'Lost, on the 9th inst., Mr. Jeremiah Hayling, aged twenty-six, a hydraulic engineer. Left his
lodgings at ten o'clock at night, and has not been heard of since. Was dressed in--'

etc., etc. Ha! That represents the last time that the colonel needed to have his machine overhauled,
I fancy."

"Good heavens!" cried my patient. "Then that explains what the girl said."

"Undoubtedly. It is quite clear that the colonel was a cool and desperate man, who was absolutely
determined that nothing should stand in the way of his little game, like those out-and-out pirates
who will leave no survivor from a captured ship. Well, every moment now is precious, so if you feel
equal to it we shall go down to Scotland Yard at once as a preliminary to starting for Eyford."

Some three hours or so afterwards we were all in the train together, bound from Reading to the
little Berkshire village. There were Sherlock Holmes, the hydraulic engineer, Inspector Bradstreet,
of Scotland Yard, a plain-clothes man, and myself. Bradstreet had spread an ordnance map of the
county out upon the seat and was busy with his compasses drawing a circle with Eyford for its
centre.