193 Holmefield Avenue
"Well, there seems to me to be only one possible solution. Mr. Rucastle seemed to be a very kind,
good-natured man. Is it not possible that his wife is a lunatic, that he desires to keep the matter
quiet for fear she should be taken to an asylum, and that he humours her fancies in every way in
order to prevent an outbreak?"

"That is a possible solution--in fact, as matters stand, it is the most probable one. But in any
case it does not seem to be a nice household for a young lady."

"But the money, Mr. Holmes, the money!"

"Well, yes, of course the pay is good--too good. That is what makes me uneasy. Why should they give
you £120 a year, when they could have their pick for £40? There must be some strong reason behind."

"I thought that if I told you the circumstances you would understand afterwards if I wanted your
help. I should feel so much stronger if I felt that you were at the back of me."

"Oh, you may carry that feeling away with you. I assure you that your little problem promises to be
the most interesting which has come my way for some months. There is something distinctly novel
about some of the features. If you should find yourself in doubt or in danger--"