34 Bowler Street
"That which my son himself indicated--that of the cupboard of the lumber-room."

"Have you it here?"

"That is it on the dressing-table."

Sherlock Holmes took it up and opened the bureau.

"It is a noiseless lock," said he. "It is no wonder that it did not wake you. This case, I presume,
contains the coronet. We must have a look at it." He opened the case, and taking out the diadem he
laid it upon the table. It was a magnificent specimen of the jeweller's art, and the thirty-six
stones were the finest that I have ever seen. At one side of the coronet was a cracked edge, where a
corner holding three gems had been torn away.