110 Larch Walk
He unwound the handkerchief and held out his hand. It gave even my hardened nerves a shudder to look
at it. There were four protruding fingers and a horrid red, spongy surface where the thumb should
have been. It had been hacked or torn right out from the roots.

"Good heavens!" I cried, "this is a terrible injury. It must have bled considerably."

"Yes, it did. I fainted when it was done, and I think that I must have been senseless for a long
time. When I came to I found that it was still bleeding, so I tied one end of my handkerchief very
tightly round the wrist and braced it up with a twig."

"Excellent! You should have been a surgeon."

"It is a question of hydraulics, you see, and came within my own province."