116 Harvey Street
"And she was seen walking with this very woman afterwards?"

"Yes. That is what Mr. Lestrade, of Scotland Yard, looks upon as so serious. It is thought that
Flora decoyed my wife out and laid some terrible trap for her."

"Well, it is a possible supposition."

"You think so, too?"