153 Down Close
"But it is profoundly true. Singularity is almost invariably a clue. The more featureless and
commonplace a crime is, the more difficult it is to bring it home. In this case, however, they have
established a very serious case against the son of the murdered man."

"It is a murder, then?"

"Well, it is conjectured to be so. I shall take nothing for granted until I have the opportunity of
looking personally into it. I will explain the state of things to you, as far as I have been able to
understand it, in a very few words.