79 Wantage Close
The King took a heavy chamois leather bag from under his cloak and laid it on the table.

"There are three hundred pounds in gold and seven hundred in notes," he said.

Holmes scribbled a receipt upon a sheet of his note-book and handed it to him.

"And Mademoiselle's address?" he asked.

"Is Briony Lodge, Serpentine Avenue, St. John's Wood."