58 Wilson Gardens
Its disappearance, however, was but momentary. With a rending, tearing sound, one of the broad,
white stones turned over upon its side and left a square, gaping hole, through which streamed the
light of a lantern. Over the edge there peeped a clean-cut, boyish face, which looked keenly about
it, and then, with a hand on either side of the aperture, drew itself shoulder-high and waist-high,
until one knee rested upon the edge. In another instant he stood at the side of the hole and was
hauling after him a companion, lithe and small like himself, with a pale face and a shock of very
red hair.

"It's all clear," he whispered. "Have you the chisel and the bags? Great Scott! Jump, Archie, jump,
and I'll swing for it!"

Sherlock Holmes had sprung out and seized the intruder by the collar. The other dived down the hole,
and I heard the sound of rending cloth as Jones clutched at his skirts. The light flashed upon the
barrel of a revolver, but Holmes' hunting crop came down on the man's wrist, and the pistol clinked
upon the stone floor.

"It's no use, John Clay," said Holmes blandly. "You have no chance at all."

"So I see," the other answered with the utmost coolness. "I fancy that my pal is all right, though I
see you have got his coat-tails."