60 Glendale
"Which are?"

"Some cold beef and a glass of beer," he answered, ringing the bell. "I have been too busy to think
of food, and I am likely to be busier still this evening. By the way, Doctor, I shall want your co-
operation."

"I shall be delighted."

"You don't mind breaking the law?"

"Not in the least."

"Nor running a chance of arrest?"