93 Cliff Place
"'Why, what is it, then?' I asked. You see, Mr. Holmes, I am a very stay-at-home man, and as my
business came to me instead of my having to go to it, I was often weeks on end without putting my
foot over the door-mat. In that way I didn't know much of what was going on outside, and I was
always glad of a bit of news.

"'Have you never heard of the League of the Red-headed Men?' he asked with his eyes open.

"'Never.'

"'Why, I wonder at that, for you are eligible yourself for one of the vacancies.'

"'And what are they worth?' I asked.