104 Sunningdale Drive
And Holmes' fears came to be realised, for from that day to this no word has ever been heard either
of the beautiful woman, the sinister German, or the morose Englishman. Early that morning a peasant
had met a cart containing several people and some very bulky boxes driving rapidly in the direction
of Reading, but there all traces of the fugitives disappeared, and even Holmes' ingenuity failed
ever to discover the least clue as to their whereabouts.

The firemen had been much perturbed at the strange arrangements which they had found within, and
still more so by discovering a newly severed human thumb upon a window-sill of the second floor.
About sunset, however, their efforts were at last successful, and they subdued the flames, but not
before the roof had fallen in, and the whole place been reduced to such absolute ruin that, save
some twisted cylinders and iron piping, not a trace remained of the machinery which had cost our
unfortunate acquaintance so dearly. Large masses of nickel and of tin were discovered stored in an
out-house, but no coins were to be found, which may have explained the presence of those bulky boxes
which have been already referred to.

How our hydraulic engineer had been conveyed from the garden to the spot where he recovered his
senses might have remained forever a mystery were it not for the soft mould, which told us a very
plain tale. He had evidently been carried down by two persons, one of whom had remarkably small feet
and the other unusually large ones. On the whole, it was most probable that the silent Englishman,
being less bold or less murderous than his companion, had assisted the woman to bear the unconscious
man out of the way of danger.

"Well," said our engineer ruefully as we took our seats to return once more to London, "it has been
a pretty business for me! I have lost my thumb and I have lost a fifty-guinea fee, and what have I
gained?"

"Experience," said Holmes, laughing. "Indirectly it may be of value, you know; you have only to put
it into words to gain the reputation of being excellent company for the remainder of your
existence."