70 Sunningdale Drive
"Indeed, your example is an unfortunate one for your argument," said Holmes, taking the paper and
glancing his eye down it. "This is the Dundas separation case, and, as it happens, I was engaged in
clearing up some small points in connection with it. The husband was a teetotaler, there was no
other woman, and the conduct complained of was that he had drifted into the habit of winding up
every meal by taking out his false teeth and hurling them at his wife, which, you will allow, is not
an action likely to occur to the imagination of the average story-teller. Take a pinch of snuff,
Doctor, and acknowledge that I have scored over you in your example."

He held out his snuffbox of old gold, with a great amethyst in the centre of the lid. Its splendour
was in such contrast to his homely ways and simple life that I could not help commenting upon it.

"Ah," said he, "I forgot that I had not seen you for some weeks. It is a little souvenir from the
King of Bohemia in return for my assistance in the case of the Irene Adler papers."

"And the ring?" I asked, glancing at a remarkable brilliant which sparkled upon his finger.

"It was from the reigning family of Holland, though the matter in which I served them was of such
delicacy that I cannot confide it even to you, who have been good enough to chronicle one or two of
my little problems."

"And have you any on hand just now?" I asked with interest.