127 Balcombe Close
"I am naturally observant, as you may have remarked, Mr. Holmes, and I soon had a pretty good plan
of the whole house in my head. There was one wing, however, which appeared not to be inhabited at
all. A door which faced that which led into the quarters of the Tollers opened into this suite, but
it was invariably locked. One day, however, as I ascended the stair, I met Mr. Rucastle coming out
through this door, his keys in his hand, and a look on his face which made him a very different
person to the round, jovial man to whom I was accustomed. His cheeks were red, his brow was all
crinkled with anger, and the veins stood out at his temples with passion. He locked the door and
hurried past me without a word or a look.

"This aroused my curiosity, so when I went out for a walk in the grounds with my charge, I strolled
round to the side from which I could see the windows of this part of the house. There were four of
them in a row, three of which were simply dirty, while the fourth was shuttered up. They were
evidently all deserted. As I strolled up and down, glancing at them occasionally, Mr. Rucastle came
out to me, looking as merry and jovial as ever.