142 Beaconside
"You are certainly joking, Holmes."

"Not in the least. Is it possible that even now, when I give you these results, you are unable to
see how they are attained?"

"I have no doubt that I am very stupid, but I must confess that I am unable to follow you. For
example, how did you deduce that this man was intellectual?"