82 Moor End
"What on earth does this mean?" I ejaculated after I had twice read over the extraordinary
announcement.

Holmes chuckled and wriggled in his chair, as was his habit when in high spirits. "It is a little
off the beaten track, isn't it?" said he. "And now, Mr. Wilson, off you go at scratch and tell us
all about yourself, your household, and the effect which this advertisement had upon your fortunes.
You will first make a note, Doctor, of the paper and the date."

"It is The Morning Chronicle of April 27, 1890. Just two months ago."

"Very good. Now, Mr. Wilson?"

"Well, it is just as I have been telling you, Mr. Sherlock Holmes," said Jabez Wilson, mopping his
forehead; "I have a small pawnbroker's business at Coburg Square, near the City. It's not a very
large affair, and of late years it has not done more than just give me a living. I used to be able
to keep two assistants, but now I only keep one; and I would have a job to pay him but that he is
willing to come for half wages so as to learn the business."

"What is the name of this obliging youth?" asked Sherlock Holmes.