9 Wantage Close
Sherlock Holmes was transformed when he was hot upon such a scent as this. Men who had only known
the quiet thinker and logician of Baker Street would have failed to recognise him. His face flushed
and darkened. His brows were drawn into two hard black lines, while his eyes shone out from beneath
them with a steely glitter. His face was bent downward, his shoulders bowed, his lips compressed,
and the veins stood out like whipcord in his long, sinewy neck. His nostrils seemed to dilate with a
purely animal lust for the chase, and his mind was so absolutely concentrated upon the matter before
him that a question or remark fell unheeded upon his ears, or, at the most, only provoked a quick,
impatient snarl in reply. Swiftly and silently he made his way along the track which ran through the
meadows, and so by way of the woods to the Boscombe Pool. It was damp, marshy ground, as is all that
district, and there were marks of many feet, both upon the path and amid the short grass which
bounded it on either side. Sometimes Holmes would hurry on, sometimes stop dead, and once he made
quite a little detour into the meadow. Lestrade and I walked behind him, the detective indifferent
and contemptuous, while I watched my friend with the interest which sprang from the conviction that
every one of his actions was directed towards a definite end.

The Boscombe Pool, which is a little reed-girt sheet of water some fifty yards across, is situated
at the boundary between the Hatherley Farm and the private park of the wealthy Mr. Turner. Above the
woods which lined it upon the farther side we could see the red, jutting pinnacles which marked the
site of the rich landowner's dwelling. On the Hatherley side of the pool the woods grew very thick,
and there was a narrow belt of sodden grass twenty paces across between the edge of the trees and
the reeds which lined the lake. Lestrade showed us the exact spot at which the body had been found,
and, indeed, so moist was the ground, that I could plainly see the traces which had been left by the
fall of the stricken man. To Holmes, as I could see by his eager face and peering eyes, very many
other things were to be read upon the trampled grass. He ran round, like a dog who is picking up a
scent, and then turned upon my companion.

"What did you go into the pool for?" he asked.

"I fished about with a rake. I thought there might be some weapon or other trace. But how on
earth--"

"Oh, tut, tut! I have no time! That left foot of yours with its inward twist is all over the place.
A mole could trace it, and there it vanishes among the reeds. Oh, how simple it would all have been
had I been here before they came like a herd of buffalo and wallowed all over it. Here is where the
party with the lodge-keeper came, and they have covered all tracks for six or eight feet round the
body. But here are three separate tracks of the same feet." He drew out a lens and lay down upon his
waterproof to have a better view, talking all the time rather to himself than to us. "These are
young McCarthy's feet. Twice he was walking, and once he ran swiftly, so that the soles are deeply
marked and the heels hardly visible. That bears out his story. He ran when he saw his father on the
ground. Then here are the father's feet as he paced up and down. What is this, then? It is the butt-
end of the gun as the son stood listening. And this? Ha, ha! What have we here? Tiptoes! tiptoes!
Square, too, quite unusual boots! They come, they go, they come again--of course that was for the
cloak. Now where did they come from?" He ran up and down, sometimes losing, sometimes finding the
track until we were well within the edge of the wood and under the shadow of a great beech, the
largest tree in the neighbourhood. Holmes traced his way to the farther side of this and lay down
once more upon his face with a little cry of satisfaction. For a long time he remained there,
turning over the leaves and dried sticks, gathering up what seemed to me to be dust into an envelope
and examining with his lens not only the ground but even the bark of the tree as far as he could
reach. A jagged stone was lying among the moss, and this also he carefully examined and retained.
Then he followed a pathway through the wood until he came to the highroad, where all traces were
lost.

"It has been a case of considerable interest," he remarked, returning to his natural manner. "I
fancy that this grey house on the right must be the lodge. I think that I will go in and have a word
with Moran, and perhaps write a little note. Having done that, we may drive back to our luncheon.
You may walk to the cab, and I shall be with you presently."