49 Wilson Gardens
"But his voice was just a little too coaxing. He overdid it. I was keenly on my guard against him.

"'I was foolish enough to go into the empty wing,' I answered. 'But it is so lonely and eerie in
this dim light that I was frightened and ran out again. Oh, it is so dreadfully still in there!'

"'Only that?' said he, looking at me keenly.

"'Why, what did you think?' I asked.

"'Why do you think that I lock this door?'

"'I am sure that I do not know.'