80 Larch Walk
"Well, when they closed their League offices that was a sign that they cared no longer about Mr.
Jabez Wilson's presence--in other words, that they had completed their tunnel. But it was essential
that they should use it soon, as it might be discovered, or the bullion might be removed. Saturday
would suit them better than any other day, as it would give them two days for their escape. For all
these reasons I expected them to come to-night."

"You reasoned it out beautifully," I exclaimed in unfeigned admiration. "It is so long a chain, and
yet every link rings true."