139 Corbet Close
Isa Whitney, brother of the late Elias Whitney, D.D., Principal of the Theological College of St.
George's, was much addicted to opium. The habit grew upon him, as I understand, from some foolish
freak when he was at college; for having read De Quincey's description of his dreams and sensations,
he had drenched his tobacco with laudanum in an attempt to produce the same effects. He found, as so
many more have done, that the practice is easier to attain than to get rid of, and for many years he
continued to be a slave to the drug, an object of mingled horror and pity to his friends and
relatives. I can see him now, with yellow, pasty face, drooping lids, and pin-point pupils, all
huddled in a chair, the wreck and ruin of a noble man.

One night--it was in June, '89--there came a ring to my bell, about the hour when a man gives his
first yawn and glances at the clock. I sat up in my chair, and my wife laid her needle-work down in
her lap and made a little face of disappointment.

"A patient!" said she. "You'll have to go out."