75 Holmefield Avenue
"Hum! So much for the police-court," said Holmes thoughtfully, tossing aside the paper. "The
question for us now to solve is the sequence of events leading from a rifled jewel-case at one end
to the crop of a goose in Tottenham Court Road at the other. You see, Watson, our little deductions
have suddenly assumed a much more important and less innocent aspect. Here is the stone; the stone
came from the goose, and the goose came from Mr. Henry Baker, the gentleman with the bad hat and all
the other characteristics with which I have bored you. So now we must set ourselves very seriously
to finding this gentleman and ascertaining what part he has played in this little mystery. To do
this, we must try the simplest means first, and these lie undoubtedly in an advertisement in all the
evening papers. If this fail, I shall have recourse to other methods."

"What will you say?"