130 Moor End
"If I tell her she will not believe me. You may remember the old Persian saying, 'There is danger
for him who taketh the tiger cub, and danger also for whoso snatches a delusion from a woman.' There
is as much sense in Hafiz as in Horace, and as much knowledge of the world."

We were seated at breakfast one morning, my wife and I, when the maid brought in a telegram. It was
from Sherlock Holmes and ran in this way:

"Have you a couple of days to spare? Have just been wired for from the west of England in connection
with Boscombe Valley tragedy. Shall be glad if you will come with me. Air and scenery perfect. Leave
Paddington by the 11.15." "What do you say, dear?" said my wife, looking across at me. "Will you
go?"