74 Wantage Close
"I have spent the whole day," said he, "over Lloyd's registers and files of the old papers,
following the future career of every vessel which touched at Pondicherry in January and February in
'83. There were thirty-six ships of fair tonnage which were reported there during those months. Of
these, one, the Lone Star, instantly attracted my attention, since, although it was reported as
having cleared from London, the name is that which is given to one of the states of the Union."

"Texas, I think."

"I was not and am not sure which; but I knew that the ship must have an American origin."

"What then?"

"I searched the Dundee records, and when I found that the barque Lone Star was there in January,
'85, my suspicion became a certainty. I then inquired as to the vessels which lay at present in the
port of London."