54 Wantage Close
"'And a very good thing, too!' I cried.

"'Yes, but you would not have me leave it a dishonoured man,' said he. 'I could not bear the
disgrace. I must raise the money in some way, and if you will not let me have it, then I must try
other means.'

"I was very angry, for this was the third demand during the month. 'You shall not have a farthing
from me,' I cried, on which he bowed and left the room without another word.