81 Cliff Place
"Only once, some years ago. I remember that it was full of papers."

"There isn't a cat in it, for example?"

"No. What a strange idea!"

"Well, look at this!" He took up a small saucer of milk which stood on the top of it.

"No; we don't keep a cat. But there is a cheetah and a baboon."

"Ah, yes, of course! Well, a cheetah is just a big cat, and yet a saucer of milk does not go very
far in satisfying its wants, I daresay. There is one point which I should wish to determine." He
squatted down in front of the wooden chair and examined the seat of it with the greatest attention.