149 Wantage Close
"You see," remarked Holmes, as we paced to and fro in front of the house, "this marriage rather
simplifies matters. The photograph becomes a double-edged weapon now. The chances are that she would
be as averse to its being seen by Mr. Godfrey Norton, as our client is to its coming to the eyes of
his princess. Now the question is--Where are we to find the photograph?"

"Where, indeed?"

"It is most unlikely that she carries it about with her. It is cabinet size. Too large for easy
concealment about a woman's dress. She knows that the King is capable of having her waylaid and
searched. Two attempts of the sort have already been made. We may take it, then, that she does not
carry it about with her."

"Where, then?"

"Her banker or her lawyer. There is that double possibility. But I am inclined to think neither.
Women are naturally secretive, and they like to do their own secreting. Why should she hand it over
to anyone else? She could trust her own guardianship, but she could not tell what indirect or
political influence might be brought to bear upon a business man. Besides, remember that she had
resolved to use it within a few days. It must be where she can lay her hands upon it. It must be in
her own house."

"But it has twice been burgled."