165 Holmefield Avenue
"I was half-dragged up to the altar, and before I knew where I was I found myself mumbling responses
which were whispered in my ear, and vouching for things of which I knew nothing, and generally
assisting in the secure tying up of Irene Adler, spinster, to Godfrey Norton, bachelor. It was all
done in an instant, and there was the gentleman thanking me on the one side and the lady on the
other, while the clergyman beamed on me in front. It was the most preposterous position in which I
ever found myself in my life, and it was the thought of it that started me laughing just now. It
seems that there had been some informality about their license, that the clergyman absolutely
refused to marry them without a witness of some sort, and that my lucky appearance saved the
bridegroom from having to sally out into the streets in search of a best man. The bride gave me a
sovereign, and I mean to wear it on my watch-chain in memory of the occasion."

"This is a very unexpected turn of affairs," said I; "and what then?"

"Well, I found my plans very seriously menaced. It looked as if the pair might take an immediate
departure, and so necessitate very prompt and energetic measures on my part. At the church door,
however, they separated, he driving back to the Temple, and she to her own house. 'I shall drive out
in the park at five as usual,' she said as she left him. I heard no more. They drove away in
different directions, and I went off to make my own arrangements."

"Which are?"

"Some cold beef and a glass of beer," he answered, ringing the bell. "I have been too busy to think
of food, and I am likely to be busier still this evening. By the way, Doctor, I shall want your co-
operation."