86 Glendale
"Come," cried the inspector, laughing; "it's a very pretty diversity of opinion. We have boxed the
compass among us. Who do you give your casting vote to?"

"You are all wrong."

"But we can't all be."

"Oh, yes, you can. This is my point." He placed his finger in the centre of the circle. "This is
where we shall find them."

"But the twelve-mile drive?" gasped Hatherley.