86 Glan Road
"Game for a morning drive?"

"Certainly."

"Then dress. No one is stirring yet, but I know where the stable-boy sleeps, and we shall soon have
the trap out." He chuckled to himself as he spoke, his eyes twinkled, and he seemed a different man
to the sombre thinker of the previous night.

As I dressed I glanced at my watch. It was no wonder that no one was stirring. It was twenty-five
minutes past four. I had hardly finished when Holmes returned with the news that the boy was putting
in the horse.

"I want to test a little theory of mine," said he, pulling on his boots. "I think, Watson, that you
are now standing in the presence of one of the most absolute fools in Europe. I deserve to be kicked
from here to Charing Cross. But I think I have the key of the affair now."