52 Wantage Close
He held out his snuffbox of old gold, with a great amethyst in the centre of the lid. Its splendour
was in such contrast to his homely ways and simple life that I could not help commenting upon it.

"Ah," said he, "I forgot that I had not seen you for some weeks. It is a little souvenir from the
King of Bohemia in return for my assistance in the case of the Irene Adler papers."

"And the ring?" I asked, glancing at a remarkable brilliant which sparkled upon his finger.

"It was from the reigning family of Holland, though the matter in which I served them was of such
delicacy that I cannot confide it even to you, who have been good enough to chronicle one or two of
my little problems."

"And have you any on hand just now?" I asked with interest.