78 Sunningdale Drive
"Which surely he restored to their owner?"

"My dear fellow, there lies the problem. It is true that 'For Mrs. Henry Baker' was printed upon a
small card which was tied to the bird's left leg, and it is also true that the initials 'H. B.' are
legible upon the lining of this hat, but as there are some thousands of Bakers, and some hundreds of
Henry Bakers in this city of ours, it is not easy to restore lost property to any one of them."