66 Moor End
"In view of your health, nothing. You are yourself aware that you will soon have to answer for your
deed at a higher court than the Assizes. I will keep your confession, and if McCarthy is condemned I
shall be forced to use it. If not, it shall never be seen by mortal eye; and your secret, whether
you be alive or dead, shall be safe with us."

"Farewell, then," said the old man solemnly. "Your own deathbeds, when they come, will be the easier
for the thought of the peace which you have given to mine." Tottering and shaking in all his giant
frame, he stumbled slowly from the room.

"God help us!" said Holmes after a long silence. "Why does fate play such tricks with poor, helpless
worms? I never hear of such a case as this that I do not think of Baxter's words, and say, 'There,
but for the grace of God, goes Sherlock Holmes.'"

James McCarthy was acquitted at the Assizes on the strength of a number of objections which had been
drawn out by Holmes and submitted to the defending counsel. Old Turner lived for seven months after
our interview, but he is now dead; and there is every prospect that the son and daughter may come to
live happily together in ignorance of the black cloud which rests upon their past.