131 Balcombe Close
"But I am so sure that he is innocent. You know what woman's instincts are. I know that he has done
no harm and that you will be sorry for having acted so harshly."

"Why is he silent, then, if he is innocent?"

"Who knows? Perhaps because he was so angry that you should suspect him."