121 Sunningdale Drive
"No, sir. But I want to find out about them, and who they are, and what their object was in playing
this prank--if it was a prank--upon me. It was a pretty expensive joke for them, for it cost them
two and thirty pounds."

"We shall endeavour to clear up these points for you. And, first, one or two questions, Mr. Wilson.
This assistant of yours who first called your attention to the advertisement--how long had he been
with you?"

"About a month then."

"How did he come?"

"In answer to an advertisement."

"Was he the only applicant?"