139 Moor End
"I have one or two things which I would wish to do now that I am in town. But I shall return by the
twelve o'clock train, so as to be there in time for your coming."

"And you may expect us early in the afternoon. I have myself some small business matters to attend
to. Will you not wait and breakfast?"