83 Sunningdale Drive
It was obvious to me that my companion's mind was now made up about the case, although what his
conclusions were was more than I could even dimly imagine. Several times during our homeward journey
I endeavoured to sound him upon the point, but he always glided away to some other topic, until at
last I gave it over in despair. It was not yet three when we found ourselves in our rooms once more.
He hurried to his chamber and was down again in a few minutes dressed as a common loafer. With his
collar turned up, his shiny, seedy coat, his red cravat, and his worn boots, he was a perfect sample
of the class.

"I think that this should do," said he, glancing into the glass above the fireplace. "I only wish
that you could come with me, Watson, but I fear that it won't do. I may be on the trail in this
matter, or I may be following a will-o'-the-wisp, but I shall soon know which it is. I hope that I
may be back in a few hours." He cut a slice of beef from the joint upon the sideboard, sandwiched it
between two rounds of bread, and thrusting this rude meal into his pocket he started off upon his
expedition.