24 Glan Road
"On the contrary, Watson, you can see everything. You fail, however, to reason from what you see.
You are too timid in drawing your inferences."

"Then, pray tell me what it is that you can infer from this hat?"

He picked it up and gazed at it in the peculiar introspective fashion which was characteristic of
him. "It is perhaps less suggestive than it might have been," he remarked, "and yet there are a few
inferences which are very distinct, and a few others which represent at least a strong balance of
probability. That the man was highly intellectual is of course obvious upon the face of it, and also
that he was fairly well-to-do within the last three years, although he has now fallen upon evil
days. He had foresight, but has less now than formerly, pointing to a moral retrogression, which,
when taken with the decline of his fortunes, seems to indicate some evil influence, probably drink,
at work upon him. This may account also for the obvious fact that his wife has ceased to love him."

"My dear Holmes!"