86 Holmefield Avenue
"Precisely so. Head attendant at the Hotel Cosmopolitan. Pray step into the cab, and I shall soon be
able to tell you everything which you would wish to know."

The little man stood glancing from one to the other of us with half-frightened, half-hopeful eyes,
as one who is not sure whether he is on the verge of a windfall or of a catastrophe. Then he stepped
into the cab, and in half an hour we were back in the sitting-room at Baker Street. Nothing had been
said during our drive, but the high, thin breathing of our new companion, and the claspings and
unclaspings of his hands, spoke of the nervous tension within him.

"Here we are!" said Holmes cheerily as we filed into the room. "The fire looks very seasonable in
this weather. You look cold, Mr. Ryder. Pray take the basket-chair. I will just put on my slippers
before we settle this little matter of yours. Now, then! You want to know what became of those
geese?"

"Yes, sir."

"Or rather, I fancy, of that goose. It was one bird, I imagine in which you were interested--white,
with a black bar across the tail."