103 Wilson Gardens
"Mr. James McCarthy, the only son of the deceased, was then called and gave evidence as follows: 'I
had been away from home for three days at Bristol, and had only just returned upon the morning of
last Monday, the 3rd. My father was absent from home at the time of my arrival, and I was informed
by the maid that he had driven over to Ross with John Cobb, the groom. Shortly after my return I
heard the wheels of his trap in the yard, and, looking out of my window, I saw him get out and walk
rapidly out of the yard, though I was not aware in which direction he was going. I then took my gun
and strolled out in the direction of the Boscombe Pool, with the intention of visiting the rabbit
warren which is upon the other side. On my way I saw William Crowder, the game-keeper, as he had
stated in his evidence; but he is mistaken in thinking that I was following my father. I had no idea
that he was in front of me. When about a hundred yards from the pool I heard a cry of "Cooee!" which
was a usual signal between my father and myself. I then hurried forward, and found him standing by
the pool. He appeared to be much surprised at seeing me and asked me rather roughly what I was doing
there. A conversation ensued which led to high words and almost to blows, for my father was a man of
a very violent temper. Seeing that his passion was becoming ungovernable, I left him and returned
towards Hatherley Farm. I had not gone more than 150 yards, however, when I heard a hideous outcry
behind me, which caused me to run back again. I found my father expiring upon the ground, with his
head terribly injured. I dropped my gun and held him in my arms, but he almost instantly expired. I
knelt beside him for some minutes, and then made my way to Mr. Turner's lodge-keeper, his house
being the nearest, to ask for assistance. I saw no one near my father when I returned, and I have no
idea how he came by his injuries. He was not a popular man, being somewhat cold and forbidding in
his manners, but he had, as far as I know, no active enemies. I know nothing further of the matter.'
"The Coroner: Did your father make any statement to you before he died? "Witness: He mumbled a few
words, but I could only catch some allusion to a rat. "The Coroner: What did you understand by that?
"Witness: It conveyed no meaning to me. I thought that he was delirious. "The Coroner: What was the
point upon which you and your father had this final quarrel? "Witness: I should prefer not to
answer. "The Coroner: I am afraid that I must press it. "Witness: It is really impossible for me to
tell you. I can assure you that it has nothing to do with the sad tragedy which followed. "The
Coroner: That is for the court to decide. I need not point out to you that your refusal to answer
will prejudice your case considerably in any future proceedings which may arise. "Witness: I must
still refuse. "The Coroner: I understand that the cry of 'Cooee' was a common signal between you and
your father? "Witness: It was. "The Coroner: How was it, then, that he uttered it before he saw you,
and before he even knew that you had returned from Bristol? "Witness (with considerable confusion):
I do not know. "A Juryman: Did you see nothing which aroused your suspicions when you returned on
hearing the cry and found your father fatally injured? "Witness: Nothing definite. "The Coroner:
What do you mean? "Witness: I was so disturbed and excited as I rushed out into the open, that I
could think of nothing except of my father. Yet I have a vague impression that as I ran forward
something lay upon the ground to the left of me. It seemed to me to be something grey in colour, a
coat of some sort, or a plaid perhaps. When I rose from my father I looked round for it, but it was
gone. "'Do you mean that it disappeared before you went for help?' "'Yes, it was gone.' "'You cannot
say what it was?' "'No, I had a feeling something was there.' "'How far from the body?' "'A dozen
yards or so.' "'And how far from the edge of the wood?' "'About the same.' "'Then if it was removed
it was while you were within a dozen yards of it?' "'Yes, but with my back towards it.' "This
concluded the examination of the witness."

"I see," said I as I glanced down the column, "that the coroner in his concluding remarks was rather
severe upon young McCarthy. He calls attention, and with reason, to the discrepancy about his father
having signalled to him before seeing him, also to his refusal to give details of his conversation
with his father, and his singular account of his father's dying words. They are all, as he remarks,
very much against the son."

Holmes laughed softly to himself and stretched himself out upon the cushioned seat. "Both you and
the coroner have been at some pains," said he, "to single out the very strongest points in the young
man's favour. Don't you see that you alternately give him credit for having too much imagination and
too little? Too little, if he could not invent a cause of quarrel which would give him the sympathy
of the jury; too much, if he evolved from his own inner consciousness anything so outré as a dying
reference to a rat, and the incident of the vanishing cloth. No, sir, I shall approach this case
from the point of view that what this young man says is true, and we shall see whither that
hypothesis will lead us. And now here is my pocket Petrarch, and not another word shall I say of
this case until we are on the scene of action. We lunch at Swindon, and I see that we shall be there
in twenty minutes."

It was nearly four o'clock when we at last, after passing through the beautiful Stroud Valley, and
over the broad gleaming Severn, found ourselves at the pretty little country-town of Ross. A lean,
ferret-like man, furtive and sly-looking, was waiting for us upon the platform. In spite of the
light brown dustcoat and leather-leggings which he wore in deference to his rustic surroundings, I
had no difficulty in recognising Lestrade, of Scotland Yard. With him we drove to the Hereford Arms
where a room had already been engaged for us.

"I have ordered a carriage," said Lestrade as we sat over a cup of tea. "I knew your energetic
nature, and that you would not be happy until you had been on the scene of the crime."